Improve prompt generation and context handling logic
This commit improves several aspects of the Synthetic Book Production workflow in the Codexes2Gemini project. Improved checks were added in the `PromptPlan.py` class for user_prompt_override, user_prompts_dict, and list_of_user_keys_to_use, ensuring that these variables are not only set but also have valid contents. A default prompt was introduced in the case where no valid prompts are provided.

Changes were made to how input context is handled in `Codexes2PartsOfTheBook.py`. It now assumes an empty string for `context_content` if `plan.context` is undefined. This reduces the likelihood of an error being introduced by unexpected null values. The `thisdoc_dir` now defaults to the current working directory if it is not set.

The ordering of import statements in `BuildLauncher.py` is improved and a new type check is added to handle differing input types for `args`. Furthermore, the logging structure was improved and now writing output in both Markdown and JSON formats. Improvements to `PartsBuilder.py` and `CodexBuilder.py` mainly consist of some aligning changes to match adjustments in other parts of the code.

These changes collectively aim to enhance robustness and error handling within the Synthetic Book Production pipeline.

Signed-off-by: Fred Zimmerman <fredzannarbor@gmail.com>