Add PartsOfTheBook classes and related CSV
Implemented a new feature which includes classes for managing the different parts of a book. Two classes, BookPart and PartsManifest, have been created to hold and list all parts of the book respectively. Two additional ones, PartsOfTheBookOrder and PartsOfTheBookLocations, are created to manage the order of the parts and their locations. Furthermore, a CSV file providing a default order for part ordering has been included.