Add README.md in userdocs directory
A new README.md file has been added to the userdocs directory. This folder is mainly provided as a convenience for users to store custom or newly created prompt and context files. Further explanation about its usage and the location of default system and user prompt files are stated in the newly created README.md file.