Add PDF generation to BuildLauncher output
Enhanced the BuildLauncher to generate a PDF output in addition to markdown and JSON formats. This is achieved using the pypandoc library, which converts markdown content to a PDF with a table of contents and specific fonts. Logging has been updated to reflect the new PDF output functionality.

If pypandoc is not installed, the application will log an error asking the user to install the library to enable PDF generation. This ensures users are informed about missing dependencies.

Signed-off-by: Fred Zimmerman <fredzannarbor@gmail.com>