Add *.pkl to .gitignore
This commit updates the .gitignore file to include the pattern "*.pkl". This ensures that any files with the .pkl extension are ignored by Git and are not included in version control. The addition helps in maintaining a cleaner repository by avoiding the accidental commit of large serialized data files, which are often binary and can clutter the repository. This change will be beneficial for developers who work with Python's pickle module, commonly used for serializing and deserializing data objects.

Signed-off-by: Fred Zimmerman <fredzannarbor@gmail.com>