Add new dependencies and bump project version to 0.2.2.0
Updated `requirements.txt` to include `google-cloud-texttospeech==2.16.5` and `pypandoc==1.13`, ensuring the project has the necessary packages for enhanced text-to-speech and document conversion functionalities. These additions are vital for the expansion of features and improving existing capabilities.

The project version in `setup.py` was incremented from 0.2.1.5 to 0.2.2.0, reflecting the integration of the new dependencies and signaling that the codebase has undergone significant updates. This version bump helps maintain proper versioning practices, ensuring clarity on the state of the project's development.

Signed-off-by: Fred Zimmerman <fredzannarbor@gmail.com>