feat: Add system_filter_submitted to PromptsPlan
This commit adds the `system_filter_submitted` boolean flag to the `PromptsPlan` class.

This flag will be used to track whether the user has submitted the system instructions filter form in the UI. This information can be useful for determining the state of the UI and controlling the flow of the application.

Signed-off-by: Fred Zimmerman <fredzannarbor@gmail.com>